+++ spaceInvaders/spaceInvaders/Headers/Enemy1.h
@@ -6,7 +6,7 @@
 class Enemy1 : public Enemy
 {
 public:
-	Enemy1(short x, short y);
+	Enemy1(float x, float y);
 	~Enemy1() = default;
 
 	void update() override;
